consolidate whois and whosent
